Created individual CSS rules for the Navbar and Dashboard
