+++ src/index.css
@@ -21,101 +21,4 @@ body {
     justify-content: center;
     min-width: 320px;
     height: 75vh;
-    background-color: #1a1a1a;
-    color: white;
-}
-
-li {
-    list-style: none;
-}
-
-a {
-    text-decoration: none;
-    color: white;
-}
-
-.navbar-card {
-    display: flex;
-    flex-direction: column;
-    gap: 24px;
-    width: 300px;
-    height: 150px;
-    border: 1px solid rgba(255, 255, 255, 0.5);
-    border-radius: 16px;
-    padding: 10px;
-    font-size: 20px;
-    background-color: hsl(249, 57%, 58%);
-    cursor: pointer;
-    transition: all 0.3s;
-}
-
-.hide {
-    display: none;
-}
-
-.navbar-card:first-child {
-    cursor: auto;
-}
-
-.navbar-card:first-child:hover {
-    transform: translateY(0px);
-}
-
-.navbar-card:hover {
-    transform: translateY(-8px);
-}
-
-.navbar-header {
-    display: flex;
-    justify-content: space-around;
-    color: #000;
-    text-align: center;
-    background-color: #fff;
-    /* display: none; */
-}
-
-.navbar-header a {
-    color: black;
-}
-
-.navbar-statistics {
-    margin-top: 22px;
-    margin-bottom: 12px;
-    font-size: 20px;
-    padding: 10px;
-    /* display: none; */
-}
-
-.navbar-card:nth-child(4) {
-    z-index: 2;
-    position: absolute;
-    bottom: -110px;
-    background-color: hwb(249 50% 14%);
-}
-
-.navbar-card:nth-child(5) {
-    z-index: 3;
-    position: absolute;
-    bottom: -220px;
-    background-color: #fff;
-}
-
-.navbar-card:nth-child(5) a {
-    color: #000;
-}
-
-nav {
-    position: relative;
-}
-
-.navbar-word-stats {
-    background-color: #1a1a1a;
-    padding: 2px 12px;
-    border-radius: 16px;
-    width: fit-content;
-}
-
-.navbar-user {
-    display: flex;
-    justify-content: space-between;
 }
